Complete revision
For one thing, we changed the table this is calling on, so that it's now using B04006, "People Reporting Ancestry." Totally wrote through the entire script.  Now includes data for the national level, the state level, the MSA level, the county level and the city/place level. Also this revision is pegged to the Russian invasion of Ukraine.